Safari-611.1.10/LayoutTests/security/mutation-observer-frame-detach-expected.txt

ALERT: test passed because nothing crashed
